+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ArtifactHelpers/FieldHelper.cs
@@ -1,21 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.SqlClient;
-using System.Linq;
-using kCura.Vendor.Castle.Core.Internal;
-using Newtonsoft.Json;
-using Relativity.Test.Helpers.ArtifactHelpers.Interfaces;
-using Relativity.Test.Helpers.Exceptions;
-using TestHelpersKepler.Interfaces.TestHelpersModule.v1.Models;
+﻿using kCura.Vendor.Castle.Core.Internal;
+using Relativity.API;
 using Relativity.Services.Interfaces.Field;
 using Relativity.Services.Interfaces.Field.Models;
 using Relativity.Services.Interfaces.Shared.Models;
-using Relativity.API;
 using Relativity.Services.Objects;
 using Relativity.Services.Objects.DataContracts;
+using Relativity.Test.Helpers.ArtifactHelpers.Interfaces;
 using Relativity.Test.Helpers.ServiceFactory.Extentions;
 using Relativity.Test.Helpers.SharedTestHelpers;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using FieldType = Relativity.Services.Interfaces.Field.Models.FieldType;
 
 namespace Relativity.Test.Helpers.ArtifactHelpers